+++ DBUtil.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.21005.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2041
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBUtil", "DBUtil.csproj", "{CADE4E0C-6240-43D1-A38E-EFBC7C542750}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBUtil", "DBUtil\DBUtil.csproj", "{81EC826A-F849-4644-A893-EE3296AD29E9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,12 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CADE4E0C-6240-43D1-A38E-EFBC7C542750}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CADE4E0C-6240-43D1-A38E-EFBC7C542750}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CADE4E0C-6240-43D1-A38E-EFBC7C542750}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CADE4E0C-6240-43D1-A38E-EFBC7C542750}.Release|Any CPU.Build.0 = Release|Any CPU
+		{81EC826A-F849-4644-A893-EE3296AD29E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{81EC826A-F849-4644-A893-EE3296AD29E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{81EC826A-F849-4644-A893-EE3296AD29E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{81EC826A-F849-4644-A893-EE3296AD29E9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CBEF0E62-C8B2-4C3E-B8D0-66267C968F97}
 	EndGlobalSection
 EndGlobal
