+++ DBUtil/PostgreSqlTableStruct.cs
