+++ DBUtil/SearchResult.cs
