+++ DBUtil/TableStruct.cs
