+++ DBUtil/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MySql.Data" version="6.9.12" targetFramework="net45" />
+  <package id="Npgsql" version="4.0.4" targetFramework="net45" />
+  <package id="Oracle.DataAccess.x86" version="2.112.1.0" targetFramework="net45" />
+  <package id="System.Data.SQLite.Core" version="1.0.109.2" targetFramework="net45" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net45" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net45" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net45" />
+</packages>
